+++ process.php
@@ -1,30 +1,18 @@
 <?php
 /**
- * Process payment
+ * Deprecated file, only for backward compatibility with old versions of the module.
+ * The process.php controller should be used instead of this file.
  */
 
-use PlacetoPay\Loggers\PaymentLogger;
+require_once 'helpers.php';
 
-try {
-    require_once 'helpers.php';
+$pathCMS = getPathCMS('process.php');
 
-    $pathCMS = getPathCMS('process.php');
+require fixPath($pathCMS . '/config/config.inc.php');
+require fixPath($pathCMS . '/init.php');
 
-    require fixPath($pathCMS . '/config/config.inc.php');
-    require fixPath($pathCMS . '/init.php');
-    require fixPath(sprintf('%s/%2$s/%2$s.php', rtrim(_PS_MODULE_DIR_, '/'), getModuleName()));
+$_GET['fc'] = 'module';
+$_GET['module'] = getModuleName();
+$_GET['controller'] = 'process';
 
-    if (!Context::getContext()->customer->isLogged()
-        && !Context::getContext()->customer->is_guest
-        && empty(file_get_contents("php://input"))) {
-        PaymentLogger::log('Access not allowed', PaymentLogger::WARNING, 17, __FILE__, __LINE__);
-
-        Tools::redirect('authentication.php?back=order.php');
-    }
-
-    (new PlacetoPayPayment())->process(isset($_GET['_']) ? $_GET['_'] : null);
-} catch (Throwable $e) {
-    PaymentLogger::log($e->getMessage(), PaymentLogger::ERROR, 999, __FILE__, __LINE__);
-
-    die($e->getMessage());
-}
+require_once dirname(__FILE__) . '/../../index.php';